Autocomplete: split up completion item provider
